+++ src/NativeImpl/LunaHook/LunaHook/emulators/vita3k.cpp
@@ -102,7 +102,7 @@ namespace
         }
         return nullptr;
     };
-    void CheckEmAddrHOOKable(uint64_t em_address, uintptr_t entrypoint)
+    void CheckEmAddrHOOKable(uint32_t em_address, uintptr_t entrypoint)
     {
         auto found = emfunctionhooks.find(em_address);
         if (found == emfunctionhooks.end())
